+++ R/deprecated.R
@@ -1,0 +1,10 @@
+#' Deprecated functions
+#' 
+#' These functions are deprecated. See the error message.
+#'
+#' @param ... parameters
+#' @rdname deprecated
+#' @export
+ggchisq_res <- function(...) {
+  stop("This function is deprecated. You could use instead JLutils::ggcross() or GGally::ggtable.")
+}