+++ NEWS.md
@@ -1,3 +1,5 @@
+# JLutils (development version)
+
 # JLutils 1.20.0
 
 * a `ggsurvey()` function for easy ggplot from a survey object